Exclude review form component and add first styling
