+++ src/modules/icmaa-review/store/getters.ts
@@ -4,19 +4,22 @@ import RootState from '@vue-storefront/core/types/RootState'
 import Review from '@vue-storefront/core/modules/review/types/Review'
 import ReviewState from '../type/ReviewState'
 
-export const arraySum = (arr): number => arr.reduce((a, b) => a + b, 0)
-export const arrayAvg = (arr): number => {
+const arraySum = (arr): number => arr.reduce((a, b) => a + b, 0)
+const arrayAvg = (arr): number => {
   const sum = arraySum(arr)
   return sum > 0 ? sum / arr.length : sum
 }
 
 const getters: GetterTree<ReviewState, RootState> = {
   getReviews: (state): Review[] => state.items.items || [],
+  getReviewAvgRating: (state) => (review: Review): number => {
+    return arrayAvg(review.ratings.map(r => r.percent))
+  },
   getReviewsCount: (state, getters): number => getters.getReviews.length || 0,
   getReviewsTotalRating: (state, getters): number => {
     return arrayAvg(
       getters.getReviews.map(
-        rvw => arrayAvg(rvw.ratings.map((r) => r.percent))
+        rvw => arrayAvg(rvw.ratings.map(r => r.percent))
       )
     )
   }
